backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
fastapi==0.109.2
ffmpeg-python==0.2.0
future==1.0.0
google-auth==2.48.0
google-genai==1.63.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.3
MarkupSafe==3.0.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==11.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.1
watchfiles==1.1.1
websockets==14.2
